split the general decoder stuff from the utils.
