+++ plonky2x/core/src/frontend/eth/rlp/utils.rs
@@ -35,54 +35,9 @@ pub struct MPTNodeFixedSize {
     pub len: usize,
 }
 
-/// An item is a string (i.e., byte array) or a list of items.
-pub enum RLPItem {
-    String(Vec<u8>),
-    List(Vec<RLPItem>),
-}
-
+// TODO:??
 // rlp_item.to_fixed_size();
 // rlp_item.to_mpt_node_fixed_size();
-
-impl Stream<u8> {
-    /// Decodes the next item in the input using RLP.
-    fn rlp_decode_next_item(&self) -> RLPItem {
-        let prefix = self.read_exact(0)[0];
-        if prefix <= 0x7F {
-            // The prefix indicates that the byte has its own RLP encoding.
-            RLPItem::String(vec![prefix])
-        } else if prefix == 0x80 {
-            // The prefix indicates this is the null value.
-            RLPItem::String(vec![])
-        } else if prefix <= 0xB7 {
-            // The prefix indicates a short string containing up to 55 bytes.
-            let length = (prefix - 0x80) as usize;
-            RLPItem::String(self.read_exact(length).to_vec())
-        } else if prefix <= 0xBF {
-            // The prefix indicates a long string containing more than 55 bytes.
-            let nb_length_bytes = (prefix - 0xB7) as usize;
-            let mut length_bytes = self.read_exact(nb_length_bytes);
-            length_bytes.reverse();
-            let mut length = 0;
-            for i in 0..nb_length_bytes {
-                length += length_bytes[i] as usize * 256_usize.pow(i as u32);
-            }
-            RLPItem::String(self.read_exact(length).to_vec())
-        } else if prefix <= 0xF7 {
-            /// The prefix indicates a short list, where the payload is 0-55 bytes.
-            let length = (prefix - 0xC0) as usize;
-            let mut elements = Vec::new();
-            for i in 0..length {
-                elements.push(self.rlp_decode_next_item());
-            }
-            RLPItem::List(elements)
-        } else {
-            // The prefix indicates a longer list.
-            let nb_length_bytes = (prefix - 0xF7) as usize;
-            todo!()
-        }
-    }
-}
 
 // pub fn rlp_decode_mpt_node(input: &[u8]) -> Vec<Vec<u8>> {
 //     info!("input {:?}", Bytes::from(input.to_vec()).to_string());
